Create 32istheAnswer!
type in 32 when the box pops up, and you for sure will get "Hello Holberton" 32 times. However many times you want to see me state that word, even once.. ten times.. your choice. In this assignment though, I would really prefer 32 because that was what was asked. I can figure a way out for 32 to be the standard number every time the code is processed as well.

Thank You